Add packages for gateways, presenters, controllers, and the load stock use case. Move StockDataLoader to gateways folder.
